micro: extract composer status bar
